requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-groq
pydantic
pypdf
bs4
langchain-huggingface
huggingface-hub
transformers
sentence-transformers
chromadb
langchain-chroma
faiss-cpu
wikipedia
streamlit
pinecone-client
langchain-pinecone
python-dotenv
pandas
plotly
